ip bans

+++ back-end/app/controllers/application_controller.rb
@@ -1,18 +1,20 @@
 class ApplicationController < ActionController::API
     def expiry_time
-        10000
+        1000
     end
     def vaildate_super token
         super_token = SuperToken.find_by(token:token)
         if super_token
-            if is_expired super_token.updated_at.to_i
-                super_token.destroy 
-                false
+            if not BannedIp.find_by(client_ip: request.remote_ip)
+                if is_expired super_token.updated_at.to_i
+                    super_token.destroy 
+                    false
+                else
+                    super_token.update(updated_at: Time.now)
+                    super_token.user
+                end
             else
-                # this might be redudant since
-                super_token.updated_at = Time.now
-                super_token.save
-                super_token.user
+                false
             end
         else
             false
@@ -21,10 +23,6 @@ class ApplicationController < ActionController::API
     def is_expired time 
         age = Time.now.to_i - time
         puts age
-        if expiry_time < age
-            true
-        else
-            false
-        end
+        expiry_time < age ? true : false
     end
 end
